Remove the PkTranslator trait
We introduced the `Translator` trait as way to combine all the functions
needed to do translation in `translate_pk` instead of passing around
functions. This made the code easier to read and easier to write.

At the same time we introduced a `PkTranslator` trait that serves the
single purpose of saving us typing out the `f_sha256` `Translator`
trait method at the cost of increasing complexity because devs must now
hold this in mind as well as `Translator`, this is a step backwards from
the initial complexity reduction improvements achieved with the
`Translator` trait.

Remove the `Translator` trait and implement `f_sha256` explicitly.