Remove f_ prefix from translator method names
We can shorten the unusual `Translator` trait method names with no loss
of meaning.

It is not immediately obvious what the `f_` prefix means, it is a relic
of history from when we used function pointers.

Remove the `f_` prefix, shortening the translator method names to just
the thing they are translating.