Enable usage of -D rustdoc::broken-intra-doc-links
Currently we have a bunch of broken links in our docs. Fix all the
broken links an add `-D rustdoc::broken-intra-doc-links` to the docs
build in CI.

There are a few other improvements scattered through this PR, I tried to
find a balance between scope creep and obvious errors in the diff.